% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_smote_cv_lasso.R
\name{par_smote_cv_lasso}
\alias{par_smote_cv_lasso}
\title{Parameter tuning for LASSO using parallel computing}
\usage{
par_smote_cv_lasso(X, y = NULL, nlambda = 100, lambda = NULL, ...)
}
\arguments{
\item{X}{feature matrix or \code{cv_smote_dataList} generated by function \code{stratified_cv_smote}}

\item{y}{binary outcome, where 1 = positive event & minority class and 0 = negative event & majority class; ignored if \code{x} is \code{cv_smote_dataList}}

\item{nlambda}{number of \code{lambda} values, default value is 100; ignored if \code{lambda} is supplied by user}

\item{lambda}{\code{lambda} sequence, typically computed by program based on \code{nlambda} and input data}

\item{\dots}{other arguments that can be passed to \code{stratified_cv_smote} or \code{glmnet}}
}
\description{
This function employs parallel computing to perform k-fold cross-validation for LASSO regression on datasets with an imbalanced binary outcome, enabling the tuning of the regularization parameter, \eqn{\lambda}.
}
\examples{
# example code

}
