% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cv.smote.lasso.R
\name{print.cv.smote.lasso}
\alias{print.cv.smote.lasso}
\title{Print the essential results for parameter tuning}
\usage{
\method{print}{cv.smote.lasso}(x, ...)
}
\arguments{
\item{x}{fitted \code{cv_smote_lasso} object}

\item{\dots}{other arguments that can be passed to \code{print}}
}
\description{
This functions prints the results of parameter tuning from cross-validation
}
