% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smote.R
\name{smote}
\alias{smote}
\title{Generate SMOTEd data ready for analysis}
\usage{
smote(X, y, k = 5, N = 9, R = 1)
}
\arguments{
\item{X}{feature matrix}

\item{y}{binary outcome, where 1 = positive event & minority class and 0 = negative event & majority class}

\item{k}{number of nearest neighbors to be considered, default value is 5}

\item{N}{number of new synthetic examples to be generated for each observation, default value is 9}

\item{R}{size ratio of the majority class to be sampled to the SMOTEd minority class, default value is 1 so that the two classes are balanced}
}
\description{
This function generates the full SMOTEd dataset with both positive and negative outcome.
}
\details{
The output dataset is consist of three parts:
1. All the original observations from the minority class (nobs = \code{sum(y)})
2. Synthetic new examples for the minority class by interpolate N values out of k nearest neighbors identified by Euclidean distance (nobs = \code{sum(y)*N})
3. Sampled observations with replacement of the majority class (nobs = \code{sum(y)\*(N+1)\*(R)})
}
\examples{
X <- matrix(rnorm(1000), ncol = 10)
y <- rbinom(100, 1, 0.1)
table(y)
smoted_data <- smote(X, y)
table(smoted_data$y)

}
