% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stratified_cv}
\alias{stratified_cv}
\title{Stratified cross-validation sampling}
\arguments{
\item{y}{binary outcome, where 1 = positive event and 0 = negative event}

\item{k}{number of folds, default value is 10}
}
\value{
A list including new synthetic examples with \code{N*nrow(X)} rows and \code{ncol(X)} columns.
}
\description{
This function assigns observations to k-fold cross validation.
}
\details{
Add details
}
