% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_cv_smote.R
\name{stratified_cv_smote}
\alias{stratified_cv_smote}
\title{Generate list of SMOTEd training data and test data for k-fold CV}
\usage{
stratified_cv_smote(X, y, k_cv = 10, k_nn = 5, N = 9, R = 1)
}
\arguments{
\item{X}{feature matrix}

\item{y}{binary outcome, where 1 = positive event & minority class and 0 = negative event & majority class}

\item{k_cv}{number of folds for the cross-validation ,default value is 10}

\item{k_nn}{number of nearest neighbors to be considered, default value is 5}

\item{N}{number of new synthetic examples to be generated for each observation, default value is 9}

\item{R}{size ratio of the majority class to be sampled to the SMOTEd minority class, default value is 1 so that the two classes are balanced}
}
\description{
This function generates a list of data sets for k-fold CV and apply SMOTE algorithm to the training set in each fold. The output will be used for parameter tuning of \eqn{\lambda} in LASSO regression.
}
\examples{
\dontrun{
X <- matrix(rnorm(1000), ncol = 10)
y <- rbinom(100, 1, 0.1)
stratified_cv_smote(X, y)
}

}
