% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cv.smote.lasso.R
\name{summary.cv.smote.lasso}
\alias{summary.cv.smote.lasso}
\title{Summarize parameter tuning results}
\usage{
\method{summary}{cv.smote.lasso}(object, s = "lambda.1se", ...)
}
\arguments{
\item{object}{fitted \code{cv_smote_lasso} object}

\item{s}{penalty parameter, default value is lambda.1se}

\item{\dots}{Additional arguments affecting the summary produced}
}
\description{
This functions prints the results of parameter tuning from cross-validation
}
