% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambda_gen}
\alias{lambda_gen}
\title{calculate the entry level of lambda}
\arguments{
\item{X}{input matrix, of dimension nobs x nvars; each row is an observation vector.}

\item{y}{binary response variable, positive outcome = 1 and negative outcome = 0.}

\item{K}{number of lambdas, by default is 100.}
}
\value{
A number
}
\description{
Calculate the smallest value for which all the beta equal to 0.
}
