% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_smote_cv_lasso.R
\name{par_smote_cv_lasso}
\alias{par_smote_cv_lasso}
\title{Parameter tuning for LASSO using parallel computing}
\usage{
par_smote_cv_lasso(X, y = NULL, nlambda = 100, lambda = NULL, ...)
}
\arguments{
\item{X}{feature matrix or \code{cv_smote_data} generated by function \code{stratified_cv_smote}}

\item{y}{binary outcome, where 1 = positive event & minority class and 0 = negative event & majority class;\
ignored if \code{x} is \code{cv_smote_data}}

\item{nlambda}{number of \code{lambda} values, default value is 100; ignored if \code{lambda} is supplied by user}

\item{lambda}{\code{lambda} sequence, typically computed by program based on \code{nlambda} and input data}

\item{\dots}{other arguments that can be passed to \code{stratified_cv_smote} or \code{glmnet}}
}
\value{
An object with class \code{"cv_smote_lasso"} including:
\item{X}{input - feature matrix}
\item{y}{input - binary outcome}
\item{lambda}{input or generated by program - lambda sequence}
\item{deviance}{matrix with \code{k} columns and \code{nlambda} rows, where each column \
contains the deviance, calculated as \eqn{-2\times \sum(ylog(p) + (1-y)log(1-p))}, for each fold}
}
\description{
This function employs parallel computing to perform k-fold cross-validation for LASSO regression \
on datasets with an imbalanced binary outcome, enabling the tuning of the regularization parameter,\
 \eqn{\lambda}.
}
\examples{
set.seed(1010)
n = 1000
p = 100
nzc = trunc(p/10)
x = matrix(rnorm(n * p), n, p)
beta = rnorm(nzc)
fx = x[, seq(nzc)] \%*\% beta
px = exp(fx)
px = px/(1 + px)
y = rbinom(n = length(px), prob = px, size = 1)
fit <- par_smote_cv_lasso(x, y, stratified = FALSE, SMOTE = FALSE)

}
