% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_smote_cv_lasso.R
\name{par_smote_cv_lasso}
\alias{par_smote_cv_lasso}
\title{Stratified cross-validation with SMOTE for LASSO using parallel computing}
\usage{
par_smote_cv_lasso(
  X,
  y = NULL,
  nlambda = 100,
  lambda = NULL,
  thresh = 1e+07,
  maxit = 1e+05,
  ...
)
}
\arguments{
\item{X}{feature matrix or \code{smote_cv_dataList} generated by function \code{stratified_cv_smote}}

\item{y}{binary outcome, where 1 = positive event & minority class and 0 = negative event & majority class; ignored if \code{x} is \code{smote_cv_dataList}.}

\item{nlambda}{number of \code{lambda} values, default value is 100.}

\item{lambda}{user supplied \code{lambda} sequence, typically computed by program based on \code{nlambda} and input data.}

\item{thresh}{convergence threshold for coordinate descent, default value is \code{1E-7}}

\item{maxit}{maximum number of passes over the data for all lambda values; default is \code{1E5}}

\item{\dots}{other arguments that can be passed to \code{stratified_cv_smote} or \code{glmnet}}
}
\description{
This function uses parallel computing to fit k-fold cross-validation for LASSO regression on datasets with an imbalanced binary outcome, where the training sets in each fold is SMOTEd, and returns a value for \eqn{\lambda}.
}
\examples{
# example code

}
