% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.glmnet.R
\name{plot.cv.glmnet}
\alias{plot.cv.glmnet}
\title{plot the cross-validation curve produced by par_smote_cv_lasso}
\usage{
\method{plot}{cv.glmnet}(fitmat)
}
\arguments{
\item{fitmat}{fitted \code{smote_cv_mseMatrix} object}
}
\description{
This functions plots the cross-validation curve, and upper and lower standard deviation curves.
}
