% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smote_minor_x}
\alias{smote_minor_x}
\title{SMOTE algorithm for minority class}
\usage{
smote_minor_x(X, k = 5L, N = 9L)
}
\arguments{
\item{X}{feature matrix}

\item{k}{number of nearest neighbors to be considered, default value is 5}

\item{N}{number of new synthetic examples to be generated for each observation, default value is 9}
}
\value{
A list including new synthetic examples with \code{N*nrow(X)} rows and \code{ncol(X)} columns.
}
\description{
This function generates synthetic examples for the minority class by kNN and interpolation
}
\details{
Add details
}
\examples{
smote_minor_x(X = matrix(rnorm(100), 10))
}
